[GN-119] feat: api schema, domain type 분리
