+++ src/types/auth.ts
@@ -1,0 +1,4 @@
+export interface TokenType {
+  refreshToken: string;
+  accessToken: string;
+}
