+++ src/types/notification.ts
@@ -1,0 +1,9 @@
+export interface MyNotificationType {
+  id: number;
+  teamId: string;
+  userId: number;
+  content: string;
+  createdAt: string;
+  updatedAt: string;
+  deletedAt: string;
+}
