+++ src/types/user.ts
@@ -1,0 +1,8 @@
+export interface UserType {
+  id: number;
+  email: string;
+  nickname: string;
+  profileImageUrl: string;
+  createdAt: string;
+  updatedAt: string;
+}
